Move slotToPos to gps.lua
