+++ autoStat.lua
@@ -1,8 +1,7 @@
-local gps = require('gps')
 local action = require('action')
 local database = require('database')
+local gps = require('gps')
 local scanner = require('scanner')
-local posUtil = require('posUtil')
 local config = require('config')
 local lowestStat
 local lowestStatSlot
@@ -58,7 +57,7 @@ local function checkChild(slot, crop)
             local stat = crop.gr + crop.ga - crop.re
 
             if stat > lowestStat then
-                action.transplant(posUtil.workingSlotToPos(slot), posUtil.workingSlotToPos(lowestStatSlot))
+                action.transplant(gps.workingSlotToPos(slot), gps.workingSlotToPos(lowestStatSlot))
                 action.placeCropStick(2)
                 database.updateFarm(lowestStatSlot, crop)
                 updateLowest()
@@ -69,7 +68,7 @@ local function checkChild(slot, crop)
             end
 
         elseif config.keepMutations and (not database.existInStorage(crop)) then
-            action.transplant(posUtil.workingSlotToPos(slot), posUtil.storageSlotToPos(database.nextStorageSlot()))
+            action.transplant(gps.workingSlotToPos(slot), gps.storageSlotToPos(database.nextStorageSlot()))
             action.placeCropStick(2)
             database.addToStorage(crop)
 
@@ -109,7 +108,7 @@ local function statOnce()
         end
 
         -- Scan
-        gps.go(posUtil.workingSlotToPos(slot))
+        gps.go(gps.workingSlotToPos(slot))
         local crop = scanner.scan()
 
         if slot % 2 == 0 then
