+++ setup.lua
@@ -6,12 +6,11 @@ local scripts = {
     'action.lua',
     'database.lua',
     'gps.lua',
-    'posUtil.lua',
     'scanner.lua',
+    'config.lua',
     'autoStat.lua',
     'autoTier.lua',
     'autoSpread.lua',
-    'config.lua',
     'uninstall.lua'
 }
